+++ packages/components/organisms/f-content-cards/vue.config.js
@@ -35,6 +35,17 @@ module.exports = {
                             ${content}`;
                 }
             });
+
+        const svgRule = config.module.rule('svg');
+
+        svgRule.uses.clear();
+
+        svgRule
+            .use('babel-loader')
+            .loader('babel-loader')
+            .end()
+            .use('vue-svg-loader')
+            .loader('vue-svg-loader');
     },
     pluginOptions: {
         lintStyleOnBuild: true